requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
ImageHash==4.0
numpy==1.18.2
Pillow==7.0.0
pkg-resources==0.0.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2019.3
PyWavelets==1.1.1
requests==2.23.0
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
